+++ src/app/core/+store/router/router.actions.ts
@@ -1,0 +1,42 @@
+import { Action } from '@ngrx/store';
+
+import { NavigationExtras } from '@angular/router';
+
+export enum RouterActionTypes {
+  GO = '[Router] GO',
+  BACK = '[Router] BACK',
+  FORWARD = '[Router] FORWARD',
+  OUTLET = '[Router] OUTLET'
+}
+
+export class Go implements Action {
+    readonly type = RouterActionTypes.GO;
+    constructor(
+        public payload: {
+            path: any[],
+            queryParams?: object,
+            extras?: NavigationExtras
+    }) { }
+}
+
+export class Back implements Action {
+    readonly type = RouterActionTypes.BACK;
+}
+
+export class Forward implements Action {
+    readonly type = RouterActionTypes.FORWARD;
+}
+
+export class Outlet implements Action {
+    readonly type = RouterActionTypes.OUTLET;
+    constructor(
+        public payload: {
+            outlets: any
+    }) { }
+}
+
+export type RouterActions
+ = Go
+ | Back
+ | Forward
+ | Outlet;