+++ src/app/core/core.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { LocalStorageService, GeneratorService, GeneratorFactory, ConstantsService } from './services';
+import { CoreStoreModule } from './+store/core-store.module';
 
 
 
@@ -8,7 +9,8 @@ import { LocalStorageService, GeneratorService, GeneratorFactory, ConstantsServi
 @NgModule({
   declarations: [],
   imports: [
-    CommonModule
+    CommonModule,
+    CoreStoreModule
   ],
   providers: [
     { provide: ConstantsService, useValue: {App: 'Shop', Version: '0.31'} },