+++ src/app/products/components/product-edit/product-edit.component.ts
@@ -1,49 +1,47 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, OnDestroy } from '@angular/core';
 
-import { ActivatedRoute, Params, Router } from '@angular/router';
-import { switchMap } from 'rxjs/operators';
-import { ProductsService } from '../../services/products.service';
 import { ProductModel } from '../../models/product';
-import { ProductsPromiseService } from '../../services/products-promise.service';
+import { Subscription } from 'rxjs';
+import { AppState } from './../../../core/+store/app.state';
+import { getSelectedProductByUrl } from './../../../core/+store/products';
+import { Store, select } from '@ngrx/store';
+import * as ProductsActions from './../../../core/+store/products/products.actions';
+import * as RouterActions from './../../../core/+store/router/router.actions';
 
 @Component({
   selector: 'app-product-edit',
   templateUrl: './product-edit.component.html',
   styleUrls: ['./product-edit.component.css']
 })
-export class ProductEditComponent implements OnInit {
+export class ProductEditComponent implements OnInit, OnDestroy {
 
-  product: ProductModel;
+  product: ProductModel = new ProductModel();
+
+  private sub: Subscription;
 
   constructor(
-    private route: ActivatedRoute,
-    private router: Router,
-    private productsService: ProductsPromiseService) {}
+    private store: Store<AppState>) {}
 
   ngOnInit() {
+    this.sub = this.store.pipe(select(getSelectedProductByUrl))
+      .subscribe(product => this.product = {...product as ProductModel});
+  }
 
-    this.product = new ProductModel();
-
-    this.route.paramMap
-      .pipe(
-        switchMap((params: Params) => this.productsService.getById(+params.get('id'))))
-      .subscribe(
-        product => this.product = {...product as ProductModel},
-        err => console.log(err)
-    );
+  ngOnDestroy() {
+    this.sub.unsubscribe();
   }
 
   onSaveProduct() {
     const product = { ...this.product };
 
     if (product.id) {
-      this.productsService.updateProduct(product).then(() => this.onGoBack());
+      this.store.dispatch(new ProductsActions.UpdateProduct(product));
     } else {
-      this.productsService.createProduct(product).then(() => this.onGoBack());
+      this.store.dispatch(new ProductsActions.CreateProduct(product));
     }
   }
 
   onGoBack(): void {
-    this.router.navigate(['/admin/products']);
+    this.store.dispatch(new RouterActions.Back());
   }
 }